requirements.txt
flask==3.0.3
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.5
transformers==4.41.2
sentence-transformers==2.7.0
torch==2.3.1
nltk==3.8.1
altair==5.3.0
pytesseract==0.3.10
Pillow==10.3.0
werkzeug==3.0.3
reportlab==4.2.2
